Refactor on_change to shared function
