+++ src/widgets/mod.rs
@@ -8,6 +8,7 @@ mod if_element;
 mod image;
 mod inspector;
 mod nine_patch;
+mod on_change;
 mod scroll;
 mod text;
 mod text_box;
@@ -24,6 +25,7 @@ pub use if_element::*;
 pub use image::*;
 pub use inspector::*;
 pub use nine_patch::*;
+pub use on_change::*;
 pub use scroll::*;
 pub use text::*;
 pub use text_box::*;
